Add the program
This is the program that utilises the data files to perform the operations.
See the README for more information. The code can be viewed in Q1.asm. The .obj file must be loaded to run the program (The .obj file is the .asm file converted into machine code (1s and 0s).